La logica funciona correcta

+++ src/Modelo/Casilla.java
@@ -4,17 +4,17 @@ package Modelo;
 public class Casilla {
  public Nodo nodo;
  public Nodo padre;
- public double g;
- public double h;
- public double f;
+ public Double g;
+ public Double h;
+ public Double f;
  public boolean abierto;
  
  public Casilla(Nodo n, boolean abierto){
 	 this.nodo = n;
 	 this.padre = null;
-	 this.g = 0;
-	 this.h = 0;
-	 this.f = 0;
+	 this.g = (double) 0;
+	 this.h = (double) 0;
+	 this.f = (double) 0;
 	 this.abierto = abierto;
  }
 }
